django_class/api_proj/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
cloudinary==1.29.0
dj3-cloudinary-storage==0.0.6
Django==4.1
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
